refactor(address-book): share credential preview routing
